Add postcss compile step
